meshlet: remove constant buffer and manual push index offset
vkCmdBindIndexBuffer supports an offset, so we used that instead of
baking it at cull time.
Since we also use the constant buffer only for output extents during
triangle culling, we moved those to push constants, freeing up some
resources and simplifying the code.

+++ src/renderer/buffer/GPUBufferView.h
@@ -13,8 +13,8 @@ struct GPUBufferProperties;
 
 struct BufferSubresource
 {
-    u32 element_offset;
-    u32 element_count;
+    u64 element_offset;
+    u64 element_count;
 };
 
 struct GPUBufferView
